requirements.txt
aiosmtplib==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
cachecontrol==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==43.0.1 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi-mail==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.114.0 ; python_version >= "3.10" and python_version < "4.0"
firebase-admin==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.19.2 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-api-python-client==2.145.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-firestore==2.18.0 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-cloud-storage==2.18.2 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
grpcio==1.66.1 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.8 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
motor==3.5.1 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==5.28.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.23.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
pymongo==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.38.5 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.30.6 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.20.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==13.0.1 ; python_version >= "3.10" and python_version < "4.0"
